refactor(client): custom-file-explorer
