+++ client/src/pages/create-project/_components/choose-template-link.tsx
@@ -159,8 +159,6 @@ export const ChooseTemplateLink = ({ sandpackTemplate }: TemplateLinkProps) => {
 	const handleOnClick = async () => {
 		if (sandpackTemplate === undefined) return console.error('ERROR');
 
-		console.log(SANDBOX_TEMPLATES[sandpackTemplate].main);
-
 		return await createProject({
 			name: 'untitled',
 			files: SANDBOX_TEMPLATES[sandpackTemplate].files,
